Move countdown to new component
